% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saucy-package.R
\docType{package}
\name{saucy-package}
\alias{saucy-package}
\title{Fast Symmetry Discovery Tool for Sparse Graphs}
\description{
Fast Symmetry Discovery Tool for Sparse Graphs
}
\author{
Bob Rudis (bob@rud.is)
}
