% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shatter.R
\name{shatter}
\alias{shatter}
\title{Identifies Symmetries in CNF Instances}
\usage{
shatter(path)
}
\arguments{
\item{path}{file with graph}
}
\description{
Identifies Symmetries in CNF Instances
}
\examples{
saucy::shatter(system.file("extdata", "example.cnf", package="saucy"))
}
\references{
\href{http://www.aloul.net/Tools/shatter/}{Shatter}
}
