% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saucy-package.R, R/saucy.R
\docType{package}
\name{saucy}
\alias{saucy}
\alias{saucy-package}
\alias{saucy}
\alias{print.saucy}
\title{Fast Symmetry Discovery Tool for Sparse Graphs}
\usage{
saucy(path, mode = c("", "cnf", "digraph", "shatter"), timeout = 0,
  rpt = 1)

\method{print}{saucy}(x, ...)
}
\arguments{
\item{path}{path}

\item{mode}{mode}

\item{timeout}{timeout}

\item{rpt}{repeat}

\item{x}{object}

\item{...}{unused}
}
\description{
Fast Symmetry Discovery Tool for Sparse Graphs

Saucy
}
\examples{
saucy::saucy(system.file("extdata", "graphfile", package="saucy"))
}
\author{
Bob Rudis (bob@rud.is)
}
