% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saucy.R
\name{saucy}
\alias{saucy}
\alias{print.saucy}
\title{Searching for Automorphisms in Underlying CNF, yes?}
\usage{
saucy(path, mode = c("", "cnf", "digraph", "shatter"), timeout = 0,
  rpt = 1)

\method{print}{saucy}(x, ...)
}
\arguments{
\item{path}{file with graph}

\item{mode}{using \code{""} auto-chooses}

\item{timeout}{timeout (Default: 0)}

\item{rpt}{repeat (mostly for benchmarking)}

\item{x}{object}

\item{...}{unused}
}
\description{
Searching for Automorphisms in Underlying CNF, yes?
}
\examples{
saucy::saucy(system.file("extdata", "graphfile", package="saucy"))
saucy::saucy(system.file("extdata", "graphfile2", package="saucy"))
}
\references{
\href{https://github.com/andandandand/saucy-repo/blob/master/saucy.c}{source}
}
